fix: inline the mixin due to problems with 3.8
